% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluate_onam}
\alias{evaluate_onam}
\title{Evaluate orthogonal neural additive model}
\usage{
evaluate_onam(model, data = model$data)
}
\arguments{
\item{model}{model of class \code{onam} as returned from \link{fit_onam} to be
evaluated}

\item{data}{Data for which the model is to be evaluated. Default is the data
with which \code{model} was fitted.}
}
\value{
Returns a list containing data, model output for each observation in
\code{data} and main and interaction effects obtained by the model
}
\description{
Evaluate orthogonal neural additive model
}
