% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_helper.R
\name{install_conda_env}
\alias{install_conda_env}
\title{Set up conda environment for keras functionality}
\usage{
install_conda_env(
  envname = "r-keras",
  python_version = "python=3.10",
  overwrite = FALSE
)
}
\arguments{
\item{envname}{Name for the conda environment to be created.}

\item{python_version}{Python version to be installed in the conda environment.}

\item{overwrite}{Should an existing conda environment of name \code{envname} be
overwritten if present?}
}
\value{
No return value, called for side effects
}
\description{
Helper function to install Keras and packages necessary for package
functionality into a conda environment. Use this function if
\code{\link[keras3:install_keras]{keras3::install_keras()}} does not work, esp. on windows machines.
}
\seealso{
\code{\link[keras3:install_keras]{keras3::install_keras()}}
}
