% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onam.R
\name{onam}
\alias{onam}
\title{Fit orthogonal neural additive model}
\usage{
onam(
  formula,
  list_of_deep_models,
  data,
  categorical_features = NULL,
  n_ensemble = 20,
  epochs = 500,
  callback = NULL,
  progresstext = FALSE,
  verbose = 0
)
}
\arguments{
\item{formula}{Formula for model fitting. Specify deep parts with the same
name as \code{list_of_deep_models}.}

\item{list_of_deep_models}{List of named models used in \code{model_formula}.}

\item{data}{Data to be fitted}

\item{categorical_features}{Vector of feature names of categorical features.}

\item{n_ensemble}{Number of orthogonal neural additive model ensembles}

\item{epochs}{Number of epochs to train the model. See
\code{\link[keras]{fit.keras.engine.training.Model}} for details.}

\item{callback}{Callback to be called during training. See
\code{\link[keras]{fit.keras.engine.training.Model}} for details.}

\item{progresstext}{Show model fitting progress. If \code{TRUE}, shows current
number of ensemble being fitted}

\item{verbose}{Verbose argument for internal model fitting. used for
debugging. See \code{\link[keras]{fit.keras.engine.training.Model}} for
details.}
}
\value{
Returns a model object of class \code{onam}, containing all ensemble
members, ensemble weights, and main and interaction effect outputs.
}
\description{
Fits an interpretable neural additive model with post hoc orthogonalization
for a given network architecture and user-specified feature sets.
}
\examples{
\donttest{
# Basic example for a simple ONAM-model
# Create training data
n <- 1000
x1 <- runif(n, -2, 2)
x2 <- runif(n, -2, 2)
y <- sin(x1) + ifelse(x2 > 0, pweibull(x2, shape = 3),
  pweibull(-x2, shape = 0.5)) +
  x1 * x2
data_train <- cbind(x1, x2, y)
# Define model
model_formula <- y ~ mod1(x1) + mod1(x2) +
  mod1(x1, x2)
list_of_deep_models <- list(mod1 = ONAM:::get_submodel)
# Fit model
callback <-
keras::keras$callbacks$EarlyStopping(monitor = "loss",
                                     patience = 10)
mod <- onam(model_formula, list_of_deep_models,
                   data_train, n_ensemble = 2, epochs = 50,
                   callback = callback,
                   progresstext = TRUE, verbose = 1)
}
}
