% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_decomp}
\alias{plot_decomp}
\title{Visualize variance decomposition}
\usage{
plot_decomp(model)
}
\arguments{
\item{model}{Model of class \code{onam} as returned from \link{onam}}
}
\value{
Returns a 'ggplot2' object showing the variance decomposition into
effects of interaction orders along with generalized sensitivity indices
}
\description{
Visualize variance decomposition
}
\examples{
\donttest{
# Basic example for a simple ONAM-model
# Create training data
n <- 1000
x1 <- runif(n, -2, 2)
x2 <- runif(n, -2, 2)
y <- sin(x1) + ifelse(x2 > 0, pweibull(x2, shape = 3),
  pweibull(-x2, shape = 0.5)) +
  x1 * x2
data_train <- cbind(x1, x2, y)
# Define model
model_formula <- y ~ mod1(x1) + mod1(x2) +
  mod1(x1, x2)
list_of_deep_models <- list(mod1 = ONAM:::get_submodel)
# Fit model
mod <- onam(model_formula, list_of_deep_models,
            data_train, n_ensemble = 2, epochs = 50,
            progresstext = TRUE, verbose = 1)
plot_decomp(mod)
}
}
