% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plotInterEffect}
\alias{plotInterEffect}
\title{Plot Main Effect}
\usage{
plotInterEffect(
  evalData,
  effect1,
  effect2,
  interpolate = FALSE,
  customColors = "spectral",
  n_interpolate = 200
)
}
\arguments{
\item{evalData}{Model output as obtained from ONAM::evaluateModel}

\item{effect1}{First effect to be plotted}

\item{effect2}{Second effect to be plotted}

\item{interpolate}{If TRUE, values will be interpolated for a smooth plot. If FALSE (default), only observations in the data will be plotted.}

\item{customColors}{color palette object for the interaction plot. Default is "spectral", returning a color palette based on the spectral theme.}

\item{n_interpolate}{number of values per coordinate axis to interpolate. Ignored if 'interpolate = FALSE'.}
}
\value{
Returns a 'ggplot2' object of the specified effect interaction
}
\description{
Plot Main Effect
}
\examples{
\donttest{
# Basic example for a simple ONAM-model
# Create training data
n <- 1000
x1 <- runif(n, -2, 2)
x2 <- runif(n, -2, 2)
y <- sin(x1) + ifelse(x2 > 0, pweibull(x2, shape = 3),
  pweibull(-x2, shape = 0.5)) +
  x1 * x2
trainDat <- cbind(x1, x2, y)
# Define model
model_formula <- y ~ mod1(x1) + mod1(x2) +
  mod1(x1, x2)
list_of_deep_models <- list(mod1 = ONAM:::getSubModel)
# Fit model
mod <- fitPHOModel(model_formula, list_of_deep_models,
                   trainDat, nEnsemble = 2,
                   progresstext = T, verbose = 1)
evalData <- evaluateModel(mod)
plotInterEffect(evalData, "x1", "x2", interpolate = TRUE)
}
}
