% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plotMainEffect}
\alias{plotMainEffect}
\title{Plot Main Effect}
\usage{
plotMainEffect(evalData, effect)
}
\arguments{
\item{evalData}{Model output as obtained from ONAM::evaluateModel}

\item{effect}{Effect to be plotted, must be present in the model formula. For interaction terms, use plotInteractionEffect}
}
\value{
Returns a ggplot2 object of the specified effect
}
\description{
Plot Main Effect
}
\examples{
\donttest{
# Basic example for a simple ONAM-model
# Create training data
n <- 1000
x1 <- runif(n, -2, 2)
x2 <- runif(n, -2, 2)
y <- sin(x1) + ifelse(x2 > 0, pweibull(x2, shape = 3),
  pweibull(-x2, shape = 0.5)) +
  x1 * x2
trainDat <- cbind(x1, x2, y)
# Define model
model_formula <- y ~ mod1(x1) + mod1(x2) +
  mod1(x1, x2)
list_of_deep_models <- list(mod1 = ONAM:::getSubModel)
# Fit model
mod <- fitPHOModel(model_formula, list_of_deep_models,
                   trainDat, nEnsemble = 2,
                   progresstext = T, verbose = 1)
evalData <- evaluateModel(mod)
plotMainEffect(evalData, "x1")
}
}
