% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onam.R
\name{summary.onam}
\alias{summary.onam}
\alias{print.summary.onam}
\title{Get summary of an onam object}
\usage{
\method{summary}{onam}(object, ...)

\method{print}{summary.onam}(x, ...)
}
\arguments{
\item{object}{onam object of class \code{onam} as returned from
\code{\link{onam}} to be summarized}

\item{...}{further arguments passed to or from other methods.}

\item{x}{object of class \code{\link{summary.onam}}.}
}
\value{
Gives summary of the \code{onam} object, including model inputs, number
of ensembles, correlation of model output and original outcome variable, and
interpretability metrics i_1 and i_2
}
\description{
generates a summary of a fitted \code{onam} object including
information on ensembling strategy and performance metrics such as
correlation and degree of interpretabiltity
}
