% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R, R/model_evaluation.R
\name{var_decomp_onam}
\alias{var_decomp_onam}
\title{Get variance decomposition of orthogonal neural additive model}
\usage{
var_decomp_onam(object, data = object$data)

var_decomp_onam(object, data = object$data)
}
\arguments{
\item{object}{Either model of class \code{onam} as returned from \link{fit_onam} or
model evaluation outcome as returned from \link{evaluate_onam}}

\item{data}{Data for which the model is to be evaluated. Default is the data
with which \code{model} was fitted.}
}
\value{
Returns a named vector of percentage of variance explained by each
interaction order.

Returns a named vector of percentage of variance explained by each
interaction order.
}
\description{
Get variance decomposition of orthogonal neural additive model

Get variance decomposition of orthogonal neural additive model
}
