% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate orthogonal neural additive model}
\usage{
evaluate_model(model_list, data = model_list$data)
}
\arguments{
\item{model_list}{model of class \code{onam} as returned from \link{fit_onam} to be
evaluated}

\item{data}{Data for which the model is to be evaluated}
}
\value{
Returns a list containing data, model output for each observation in
\code{data} and main and interaction effects obtained by the model
}
\description{
Evaluate orthogonal neural additive model
}
