% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_inter_effect}
\alias{plot_inter_effect}
\title{Plot Interaction Effect}
\usage{
plot_inter_effect(
  data_eval,
  effect1,
  effect2,
  interpolate = FALSE,
  custom_colors = "spectral",
  n_interpolate = 200
)
}
\arguments{
\item{data_eval}{Model output as obtained from ONAM::evaluate_model}

\item{effect1}{First effect to be plotted}

\item{effect2}{Second effect to be plotted}

\item{interpolate}{If TRUE, values will be interpolated for a smooth plot.
If FALSE (default), only observations in the data will be plotted.}

\item{custom_colors}{color palette object for the interaction plot. Default
is "spectral", returning a color palette based on the spectral theme.}

\item{n_interpolate}{number of values per coordinate axis to interpolate.
Ignored if 'interpolate = FALSE'.}
}
\value{
Returns a 'ggplot2' object of the specified effect interaction
}
\description{
Plot Interaction Effect
}
\examples{
\donttest{
# Basic example for a simple ONAM-model
# Create training data
n <- 1000
x1 <- runif(n, -2, 2)
x2 <- runif(n, -2, 2)
y <- sin(x1) + ifelse(x2 > 0, pweibull(x2, shape = 3),
  pweibull(-x2, shape = 0.5)) +
  x1 * x2
data_train <- cbind(x1, x2, y)
# Define model
model_formula <- y ~ mod1(x1) + mod1(x2) +
  mod1(x1, x2)
list_of_deep_models <- list(mod1 = ONAM:::get_submodel)
# Fit model
mod <- fit_onam(model_formula, list_of_deep_models,
                   data_train, n_ensemble = 2, epochs = 50,
                   progresstext = TRUE, verbose = 1)
data_eval <- evaluate_model(mod)
plot_inter_effect(data_eval, "x1", "x2", interpolate = TRUE)
}
}
