% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\docType{methods}
\name{predict.onam}
\alias{predict.onam}
\alias{predict}
\title{Evaluate orthogonal neural additive model}
\usage{
\method{predict}{onam}(object, ..., data = NULL)
}
\arguments{
\item{object}{model of class \code{onam} as returned from \link{onam} to be
evaluated}

\item{...}{some methods for this generic require additional arguments. None
are used in this method.}

\item{data}{Data for which the model is to be evaluated. If NULL (default),
data with which \code{model} was fitted is used.}
}
\value{
Returns a list containing data, model output for each observation in
\code{data} and main and interaction effects obtained by the model
}
\description{
Evaluate orthogonal neural additive model
}
