% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{var_decomp_onam}
\alias{var_decomp_onam}
\title{Get variance decomposition of orthogonal neural additive model}
\usage{
var_decomp_onam(object, data = NULL)
}
\arguments{
\item{object}{Either model of class \code{onam} as returned from \link{fit_onam} or
model evaluation outcome as returned from \link{evaluate_onam}}

\item{data}{Data for which the model is to be evaluated. If \code{NULL}
(DEFAULT), the data from model fitting is used.
with which \code{model} was fitted.}
}
\value{
Returns a named vector of percentage of variance explained by each
interaction order.
}
\description{
Get variance decomposition of orthogonal neural additive model
}
\examples{
\donttest{
# Basic example for a simple ONAM-model
# Create training data
n <- 1000
x1 <- runif(n, -2, 2)
x2 <- runif(n, -2, 2)
y <- sin(x1) + ifelse(x2 > 0, pweibull(x2, shape = 3),
  pweibull(-x2, shape = 0.5)) +
  x1 * x2
data_train <- cbind(x1, x2, y)
# Define model
model_formula <- y ~ mod1(x1) + mod1(x2) +
  mod1(x1, x2)
list_of_deep_models <- list(mod1 = ONAM:::get_submodel)
# Fit model
mod <- onam(model_formula, list_of_deep_models,
            data_train, n_ensemble = 2, epochs = 50,
            progresstext = TRUE, verbose = 1)
var_decomp_onam(mod)
}
}
